pyproject.toml
[project]
name = "dteam-django-practical-test"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "pdfkit (>=1.0.0,<2.0.0)",
    "ipython (>=9.3.0,<10.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "openai (>=1.90.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
